renaming to pubmetric
