+++ src/pubmetric/api_controller.py
@@ -6,9 +6,9 @@ import os
 import json
 from datetime import datetime
 
-from wfqc.metrics import *
-from wfqc.workflow import parse_cwl_workflows
-from wfqc.network import create_citation_network
+from pubmetric.metrics import *
+from pubmetric.workflow import parse_cwl_workflows
+from pubmetric.network import create_citation_network
 
 app = FastAPI()
